+++ packages/tooltip-react/src/Tooltip.tsx
@@ -51,7 +51,7 @@ export const Tooltip = ({ content, initialPlacement = "top", className }: Toolti
     ]);
 
     return (
-        <span className={cn("jkl-tooltip", className)} aria-live="polite">
+        <span className={cn("jkl-tooltip", className)}>
             <span className="jkl-tooltip__wrapper">
                 <button type="button" className="jkl-tooltip__button" {...getReferenceProps({ ref: reference })}>
                     <span aria-hidden="true">?</span>